Click submenu opening control: toggle all checkbox
Marked with: INS251229 Baertram 
At the top of the file, and at the context menu of antiquities and survey filters

Clicking the submenu opening controls toggles all checkboxes in the submenu to checked/unchecked
e.g. clicking "Antiquties" will toggle the 3 checkboxes for leads etc. in the submenu to "all on" or "all off". No extra SavedVariables used for that.
Default value will be "All off" so first click will toggle all "ON" and next all "OFF".